moved validation method to language service
