+++ packages/serverless-workflow-language-service/package.json
@@ -28,7 +28,7 @@
     "vscode-languageserver-types": "^3.16.0",
     "jsonc-parser": "^3.0.0",
     "vscode-languageserver-textdocument": "^1.0.4",
-    "vscode-json-languageservice": "4.2.1"
+    "vscode-json-languageservice": "^4.2.1"
   },
   "babel": {
     "presets": [
