+++ packages/serverless-workflow-language-service/src/channel/index.ts
@@ -1,2 +1,1 @@
 export * from "./SwfJsonLanguageService";
-export * from "./SwfJsonValidation";
